+++ src/queue/PruneVolumeBackupsJob.php
@@ -10,7 +10,7 @@ class PruneVolumeBackupsJob extends BaseJob
 {
     public function execute($queue)
     {
-        RemoteBackup::getInstance()->remotebackup->pruneVolumes();
+        RemoteBackup::getInstance()->pruneservice->pruneVolumes();
     }
 
     protected function defaultDescription()